幼儿管理的增删改查 家长管理的增删改查

+++ src/main/java/com/great/childschool/entity/CcTblParentBaby.java
@@ -4,45 +4,49 @@ package com.great.childschool.entity;
 public class CcTblParentBaby
 {
 
-  private long pbid;
-  private long pid;
-  private long bid;
+  private int pbid;
+  private int pid;
+  private int  bid;
   private String prelation;
 
 
-  public long getPbid() {
+  public int getPbid()
+  {
     return pbid;
   }
 
-  public void setPbid(long pbid) {
+  public void setPbid(int pbid)
+  {
     this.pbid = pbid;
   }
 
-
-  public long getPid() {
+  public int getPid()
+  {
     return pid;
   }
 
-  public void setPid(long pid) {
+  public void setPid(int pid)
+  {
     this.pid = pid;
   }
 
-
-  public long getBid() {
+  public int getBid()
+  {
     return bid;
   }
 
-  public void setBid(long bid) {
+  public void setBid(int bid)
+  {
     this.bid = bid;
   }
 
-
-  public String getPrelation() {
+  public String getPrelation()
+  {
     return prelation;
   }
 
-  public void setPrelation(String prelation) {
+  public void setPrelation(String prelation)
+  {
     this.prelation = prelation;
   }
-
 }
